+++ index.js
@@ -88,8 +88,30 @@ class Jetpack {
     this.serverless = serverless;
     this.options = options;
 
+    this.commands = {
+      jetpack: {
+        usage: "Alternate Serverless packager",
+        commands: {
+          "package": {
+            usage: "Packages a Serverless service or function",
+            lifecycleEvents: [
+              "package"
+            ],
+            options: {
+              "function": {
+                usage: "Function name. Packages a single function (see 'deploy function')",
+                shortcut: "f"
+              }
+            }
+          }
+        }
+      }
+    };
+
     this.hooks = {
-      "before:package:createDeploymentArtifacts": this.package.bind(this)
+      "before:package:createDeploymentArtifacts": this.package.bind(this),
+      "before:package:function:package": this.package.bind(this),
+      "jetpack:package:package": this.package.bind(this)
     };
   }
 
@@ -379,8 +401,17 @@ class Jetpack {
     const { service } = this.serverless;
     const servicePackage = service.package;
 
+    // Check if we have a single function limitation from `deploy -f name`.
+    const singleFunctionName = (this.options || {}).function;
+    if (singleFunctionName) {
+      this._logDebug(`Packaging only for function: ${singleFunctionName}`);
+    }
+
     // Gather internal configuration.
     const fnsPkgs = service.getAllFunctions()
+      // Limit to single function if provided.
+      .filter((functionName) => !singleFunctionName || singleFunctionName === functionName)
+      // Convert to more useful format.
       .map((functionName) => ({
         functionName,
         functionObject: service.getFunction(functionName)
@@ -396,6 +427,7 @@ class Jetpack {
         artifact: obj.functionPackage.artifact
       }));
 
+    // Get list of individual functions to package.
     const fnsPkgsToPackage = fnsPkgs.filter((obj) =>
       (servicePackage.individually || obj.individually) && !(obj.disable || obj.artifact)
     );
@@ -410,10 +442,14 @@ class Jetpack {
     // We recreate the logic from `packager#packageService` for deciding whether
     // to package the service or not.
     const shouldPackageService = !servicePackage.individually
+      && !servicePackage.artifact
+      // Don't package service if we specify a single function **and** have a match
+      && (!singleFunctionName || !fnsPkgsToPackage.length)
+      // Otherwise, have some functions left that need to use the service package.
       && fnsPkgs.some((obj) => !(obj.disable || obj.individually || obj.artifact));
 
     // Package entire service if applicable.
-    if (shouldPackageService && !servicePackage.artifact) {
+    if (shouldPackageService) {
       await this.packageService();
     } else if (!fnsPkgsToPackage.length) {
       // Detect if we did nothing...